Moved if func into its own file, added tests
